requirements.txt
torch
torchvision
opencv-python
tqdm
pyyaml
scikit-learn
matplotlib
pillow